issue #169 refactor: directory change
