+++ src/components/Header/SubHeader/SubHeader.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import Tab from "@/components/Base/Tab/Tab";
+import Tab from "@/components/Common/Tab/Tab";
 
 import { SubHeaderDiv } from "./SubHeader.style";
